Moves listingDetails2 into its own folder with its own structure. Also seperates the imageGallery to its own standalone component.
